requirements.txt
streamlit>=1.38.0
PyPDF2>=3.0.1
python-docx>=1.0.1
pandas>=2.2.3
openai>=1.50.0
